feat: rename metrics to observability
