+++ coordinator/internal/controller/api/controller.go
@@ -17,10 +17,6 @@ var (
 	SubmitProof *SubmitProofController
 	// Auth the auth controller
 	Auth *AuthController
-	// HealthCheck the health check controller
-	HealthCheck *HealthCheckController
-	// Ready the ready controller
-	Ready *ReadyController
 
 	initControllerOnce sync.Once
 )
@@ -34,9 +30,7 @@ func InitController(cfg *config.Config, db *gorm.DB, reg prometheus.Registerer) 
 		}
 
 		Auth = NewAuthController(db)
-		HealthCheck = NewHealthCheckController(db)
 		GetTask = NewGetTaskController(cfg, db, vf, reg)
 		SubmitProof = NewSubmitProofController(cfg, db, vf, reg)
-		Ready = NewReadyController()
 	})
 }
