+++ prover-stats-api/internal/route/route.go
@@ -2,8 +2,11 @@ package route
 
 import (
 	"github.com/gin-gonic/gin"
+	"github.com/prometheus/client_golang/prometheus"
 	swaggerFiles "github.com/swaggo/files"
 	ginSwagger "github.com/swaggo/gin-swagger"
+
+	"scroll-tech/common/observability"
 
 	"scroll-tech/prover-stats-api/internal/config"
 	"scroll-tech/prover-stats-api/internal/controller"
@@ -11,11 +14,10 @@ import (
 )
 
 // Route routes the APIs
-func Route(router *gin.Engine, conf *config.Config) {
+func Route(router *gin.Engine, conf *config.Config, reg prometheus.Registerer) {
 	router.GET("/swagger/*any", ginSwagger.WrapHandler(swaggerFiles.Handler))
 
-	router.GET("api/health", controller.HealthCheck.HealthCheck)
-	router.GET("api/ready", controller.Ready.Ready)
+	observability.Use(router, "prover_stats_api", reg)
 
 	r := router.Group("api/prover_task")
